collector/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
isort==5.10.1
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
